+++ packages/terra-application-navigation/CHANGELOG.md
@@ -8,6 +8,7 @@ Unreleased
 
 ### Changed
 * Updated `content-layout` background color.
+* Fixed `settings` api called when `help` clicked in drawer menu.
 
 1.18.0 - (February 4, 2020)
 ------------------