+++ packages/terra-application-navigation/CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+
+1.19.0 - (February 11, 2020)
+------------------
 ### Fixed
 * Fixed `settings` api called when `help` clicked in drawer menu.